Add memories_add_enriched for AI-powered knowledge creation
- Add EnrichmentParams struct and memories_add_enriched() to memories.rs
- Refactor tool_create_knowledge.rs to use memories_add_enriched (254→93 lines)
- Add enriched memory creation to tool_deep_research.rs
- Add enriched memory creation to tool_strategic_planning.rs
- All knowledge entries now get AI-enriched metadata and auto-deprecation
