+++ refact-agent/engine/src/tools/tool_create_knowledge.rs
@@ -1,36 +1,17 @@
 use std::collections::HashMap;
 use std::sync::Arc;
 use async_trait::async_trait;
-use regex::Regex;
 use serde_json::Value;
-use tracing::{info, warn};
+use tracing::info;
 use tokio::sync::Mutex as AMutex;
 
 use crate::at_commands::at_commands::AtCommandsContext;
 use crate::call_validation::{ChatMessage, ChatContent, ContextEnum};
 use crate::tools::tools_description::{Tool, ToolDesc, ToolParam, ToolSource, ToolSourceType};
-use crate::knowledge_graph::kg_subchat::{enrich_knowledge_metadata, check_deprecation};
-use crate::knowledge_graph::kg_structs::KnowledgeFrontmatter;
-use crate::memories::{memories_add, deprecate_document};
-use crate::global_context::try_load_caps_quickly_if_not_present;
+use crate::memories::{memories_add_enriched, EnrichmentParams};
 
 pub struct ToolCreateKnowledge {
     pub config_path: String,
-}
-
-fn extract_entities(content: &str) -> Vec<String> {
-    let backtick_re = Regex::new(r"`([a-zA-Z_][a-zA-Z0-9_:]*(?:::[a-zA-Z_][a-zA-Z0-9_]*)*)`").unwrap();
-    backtick_re.captures_iter(content)
-        .map(|c| c.get(1).unwrap().as_str().to_string())
-        .filter(|e| e.len() >= 3 && e.len() <= 100)
-        .collect()
-}
-
-fn extract_file_paths(content: &str) -> Vec<String> {
-    let path_re = Regex::new(r"(?:^|[\s`])((?:[a-zA-Z0-9_-]+/)+[a-zA-Z0-9_-]+\.[a-zA-Z0-9]+)").unwrap();
-    path_re.captures_iter(content)
-        .map(|c| c.get(1).unwrap().as_str().to_string())
-        .collect()
 }
 
 #[async_trait]
@@ -77,8 +58,6 @@ impl Tool for ToolCreateKnowledge {
     ) -> Result<(bool, Vec<ContextEnum>), String> {
         info!("create_knowledge {:?}", args);
 
-        let gcx = ccx.lock().await.global_context.clone();
-
         let content = match args.get("content") {
             Some(Value::String(s)) => s.clone(),
             Some(v) => return Err(format!("argument `content` is not a string: {:?}", v)),
@@ -95,149 +74,16 @@ impl Tool for ToolCreateKnowledge {
             _ => vec![],
         };
 
-        let entities = extract_entities(&content);
-        let detected_paths = extract_file_paths(&content);
-
-        let caps = try_load_caps_quickly_if_not_present(gcx.clone(), 0).await
-            .map_err(|e| format!("Failed to load caps: {}", e.message))?;
-        let light_model = if caps.defaults.chat_light_model.is_empty() {
-            caps.defaults.chat_default_model.clone()
-        } else {
-            caps.defaults.chat_light_model.clone()
+        let enrichment_params = EnrichmentParams {
+            base_tags: user_tags,
+            base_filenames: user_filenames,
+            base_kind: "knowledge".to_string(),
+            base_title: None,
         };
 
-        let kg = crate::knowledge_graph::build_knowledge_graph(gcx.clone()).await;
+        let file_path = memories_add_enriched(ccx.clone(), &content, enrichment_params).await?;
 
-        let candidate_files: Vec<String> = {
-            let mut files = user_filenames.clone();
-            files.extend(detected_paths);
-            files.into_iter().take(30).collect()
-        };
-
-        let candidate_docs: Vec<(String, String)> = kg.active_docs()
-            .take(20)
-            .map(|d| {
-                let id = d.frontmatter.id.clone().unwrap_or_else(|| d.path.to_string_lossy().to_string());
-                let title = d.frontmatter.title.clone().unwrap_or_else(|| "Untitled".to_string());
-                (id, title)
-            })
-            .collect();
-
-        let enrichment = enrich_knowledge_metadata(
-            ccx.clone(),
-            &light_model,
-            &content,
-            &entities,
-            &candidate_files,
-            &candidate_docs,
-        ).await;
-
-        let (final_title, final_tags, final_filenames, final_kind, final_links, review_days) = match enrichment {
-            Ok(e) => {
-                let mut tags = user_tags.clone();
-                tags.extend(e.tags);
-                tags.sort();
-                tags.dedup();
-
-                let mut files = user_filenames.clone();
-                files.extend(e.filenames);
-                files.sort();
-                files.dedup();
-
-                let kind = e.kind.unwrap_or_else(|| if files.is_empty() { "domain".to_string() } else { "code".to_string() });
-
-                (
-                    e.title.or_else(|| content.lines().next().map(|l| l.trim_start_matches('#').trim().to_string())),
-                    if tags.is_empty() { vec!["knowledge".to_string()] } else { tags },
-                    files,
-                    kind,
-                    e.links,
-                    e.review_after_days.unwrap_or(90),
-                )
-            }
-            Err(e) => {
-                warn!("Enrichment failed, using defaults: {}", e);
-                let tags = if user_tags.is_empty() { vec!["knowledge".to_string()] } else { user_tags };
-                let kind = if user_filenames.is_empty() { "domain".to_string() } else { "code".to_string() };
-                (
-                    content.lines().next().map(|l| l.trim_start_matches('#').trim().to_string()),
-                    tags,
-                    user_filenames,
-                    kind,
-                    vec![],
-                    90,
-                )
-            }
-        };
-
-        let now = chrono::Local::now();
-        let frontmatter = KnowledgeFrontmatter {
-            id: Some(uuid::Uuid::new_v4().to_string()),
-            title: final_title.clone(),
-            tags: final_tags.clone(),
-            created: Some(now.format("%Y-%m-%d").to_string()),
-            updated: Some(now.format("%Y-%m-%d").to_string()),
-            filenames: final_filenames.clone(),
-            links: final_links,
-            kind: Some(final_kind.clone()),
-            status: Some("active".to_string()),
-            superseded_by: None,
-            deprecated_at: None,
-            review_after: Some((now + chrono::Duration::days(review_days)).format("%Y-%m-%d").to_string()),
-        };
-
-        let file_path = memories_add(gcx.clone(), &frontmatter, &content).await?;
-        let new_doc_id = frontmatter.id.clone().unwrap();
-
-        let deprecation_candidates = kg.get_deprecation_candidates(
-            &final_tags,
-            &final_filenames,
-            &entities,
-            Some(&new_doc_id),
-        );
-
-        let mut deprecated_count = 0;
-        if !deprecation_candidates.is_empty() {
-            let snippet: String = content.chars().take(500).collect();
-
-            match check_deprecation(
-                ccx.clone(),
-                &light_model,
-                final_title.as_deref().unwrap_or("Untitled"),
-                &final_tags,
-                &final_filenames,
-                &snippet,
-                &deprecation_candidates,
-            ).await {
-                Ok(result) => {
-                    for decision in result.deprecate {
-                        if decision.confidence >= 0.75 {
-                            if let Some(doc) = kg.get_doc_by_id(&decision.target_id) {
-                                if let Err(e) = deprecate_document(
-                                    gcx.clone(),
-                                    &doc.path,
-                                    Some(&new_doc_id),
-                                    &decision.reason,
-                                ).await {
-                                    warn!("Failed to deprecate {}: {}", decision.target_id, e);
-                                } else {
-                                    deprecated_count += 1;
-                                    info!("Deprecated {} (confidence: {:.2}): {}", decision.target_id, decision.confidence, decision.reason);
-                                }
-                            }
-                        }
-                    }
-                }
-                Err(e) => {
-                    warn!("Deprecation check failed: {}", e);
-                }
-            }
-        }
-
-        let mut result_msg = format!("Knowledge entry created: {}", file_path.display());
-        if deprecated_count > 0 {
-            result_msg.push_str(&format!("\n{} outdated document(s) marked as deprecated.", deprecated_count));
-        }
+        let result_msg = format!("Knowledge entry created: {}", file_path.display());
 
         Ok((false, vec![ContextEnum::ChatMessage(ChatMessage {
             role: "tool".to_string(),
